Update of guide to reflect changes in our system
Included information about the CRITICAL importance of using the proper API server, along with a bunch of changes to language that were causing huge problems for people trying to use this guide.

I STRONGLY suggest someone on the team looks at the Korean version and updates it following this. Seriously.